sync: legacy rust/dev v0.4.15 -> staging (+ re-applied polish)
Squashed onto staging from sync/legacy-v0.4.15 because the source branch is
rooted in legacy's history (unrelated to staging's). Reference branch with
the 6-commit logical structure: sync/legacy-v0.4.15.

The 6 logical commits this squash represents:

1. chore: strip legacy Scala/JVM artifacts from base (~1340 files: nix/,
   project/, rosette/, rspace/, rspace-bench/, *.scala/*.sbt, /scala/,
   /java/, /scalapb/, /debian/, /rpm/, /slowcooker/, per-crate Cargo.lock,
   legacy .github/)

2. chore: import new-repo native infrastructure on top of legacy
   (.github/workflows/ci.yml, .githooks/, AGENTS/DEVELOPER/GEMINI.md,
   deny.toml, rustfmt.toml; replace docs/docker/scripts/examples with new
   repo's versions; preserve stigmergic docs, OCIR + VPS deploy infra;
   keep legacy's lean post-#492 docker/conf/{default,standalone-dev}.conf;
   drop docker/helm/.../logback.xml, kamon files, docker/.env)

3. chore: reconcile root files (take new repo's .gitignore, CLAUDE.md,
   CONTRIBUTING.md, README.md, coding_guidelines.md, rust-toolchain.toml;
   keep legacy's Cargo.toml (metrics 0.24 bump), Justfile (superset),
   CHANGELOG.md (v0.4.13-v0.4.15 entries))

4. chore: re-apply crate-level clippy allow blocks (polish group 3a) --
   #![allow(clippy::...)] re-prepended to 10 lib.rs files; restore
   scalapb/scalapb.proto shims needed by protoc

5. chore: re-apply targeted clippy polish + docker image rename --
   blake2b256_hash.rs allow, deploy_chain_index.rs allow, docker-commands.sh
   rename f1r3fly-rust-node -> f1r3node-rust

6. deps(lmdb): upgrade heed 0.11.0 -> 0.22.1 + shared env cache (polish
   group 2) -- adapt LMDB code to heed 0.22 API; new env_cache.rs
   (shared Weak<Env> cache prevents EnvAlreadyOpened aliasing); LMDB test
   rewrites; Cargo.lock regenerated

Captures legacy v0.4.13 -> v0.4.15 delta (PRs #488, #491, #492, #497, #499,
#502, #503, #504, #505): BitmaskOR mergeable channels (foundation for D-thin
Shape A), self-contained binary (#492 embeds Rholang resources + defaults.conf,
eliminates kamon), lock-free RuntimeManager, ChromaDB integration,
deploy_finalization_status API edge cases, merge-stale-diff bug class fix.

Verified: cargo test --workspace --release --no-fail-fast on sync branch:
2184 tests, 2147 passed, 0 failed, 37 ignored (vs baseline 2106/2075/0/31
on origin/staging -- +78 imported tests, zero regressions).

Follow-up work documented in docs/sync-todo.md (markdown link checker, FFI
removal, VPS config unification, ScalaPB removal, contributor-surface
refresh, CI hardening).

+++ rholang/src/rust/interpreter/registry/registry.rs
@@ -24,9 +24,9 @@ impl CRC14 {
         for _ in 0..8 {
             let shift_rem = rem << 1;
             if (shift_rem & 0x4000) != 0 {
-                rem = shift_rem ^ 0x4805;
+                rem = (shift_rem ^ 0x4805) as u16;
             } else {
-                rem = shift_rem;
+                rem = shift_rem as u16;
             }
         }
